Getting replication lag from production database
We used to select from repl_mon in postgres DB. This patch changes it to
select from the same DB that is used for user queries because:
1. Regexp was designed so that dbname is a single word (i.e. you could
not use dash in dbname). So it makes really difficult to use pgbouncer
on the same host as PL/Proxy to save connections.
2. We now have repl_mon table in every DB through FDW to postgres DB.